+++ Demo/Keyboard/DemoKeyboardActionHandler.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardActionHandler.swift
-//  KeyboardKit
+//  Keyboard
 //
 //  Created by Daniel Saidi on 2021-02-11.
 //  Copyright © 2021-2023 Daniel Saidi. All rights reserved.