+++ Demo/KeyboardCustom/DemoKeyboardAppearance.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardAppearance.swift
-//  Keyboard
+//  KeyboardCustom
 //
 //  Created by Daniel Saidi on 2022-09-02.
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.