+++ Demo/KeyboardCustom/DemoKeyboardView.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardView.swift
-//  KeyboardKitDemo
+//  KeyboardCustom
 //
 //  Created by Daniel Saidi on 2022-02-04.
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
@@ -10,17 +10,13 @@ import KeyboardKit
 import SwiftUI
 
 /**
- This view wraps a `SystemKeyboard`, which aims to mimic the
- native iOS keyboard as closely as possible.
+ This view adds a `SystemKeyboard` with some more views to a
+ `VStack`, then configures the keyboard slightly.
 
- While the `SystemKeyboard` is very flexible, it can also be
- easily created by just providing it with a controller, like
- we do here. This configuration adds an autocomplete toolbar
- above the keyboard and replaces the system keyboard with an
- emoji keyboard whenever needed.
-
- This view adds some extra views around the keyboard to show
- how easy it is to use native SwiftUI with KeyboardKit.
+ The system keyboard will then adapt its appearance based on
+ the ``DemoKeyboardAppearance``, its input keys based on the
+ ``DemoInputSetProvider`` and its full keyboard layout based
+ on the ``DemoLayoutProvider``.
  */
 struct DemoKeyboardView: View {