+++ Demo/KeyboardPro/DemoKeyboardLayoutProvider.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardLayoutProvider.swift
-//  Keyboard
+//  KeyboardPro
 //
 //  Created by Daniel Saidi on 2022-12-21.
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.