+++ Demo/KeyboardPro/DemoKeyboardView.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardView.swift
-//  KeyboardUnicode
+//  KeyboardPro
 //
 //  Created by Daniel Saidi on 2022-02-04.
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.